프롬프트: 몇가지 수정사항이 더 있어. 1. 캘린더가 어두워서 똑같이 어두운 숫자가 잘 안보이는데, 숫자를 밝은 색으로 바꿔줘 2. 촬영을 끝내거나 창을 닫으면 미리 설정해둔 동영상 설정에 따라 자동으로 영상을 만들고 저장하게 해줘 3. 1초 내지는 그 이하의 간격으로 촬영을 할건데, 그럼 카메라를 자원 해체했다 가져오는 걸 반복하는 게 더 비효율적이야. 그리니까 카메라를 계속 가져오게 바꾸고, 대신 10프레임 정도의 fps로 호출하게 바꿔봐. 4. 캘린더에서 플래너 열기를 하면 카메라가 있는 창까지 열려. 플래너를 볼 수만 있는(추가도 못하는) 창을 따로 만들어서 그 창에서 열리게 해줘
그리고 requirements.txt 대신 pyproject.toml로, pyproject.toml 내용도 제대로 작성해서, 깃허브로 설치할 수 있게 라이브러리화 해줘. 깃허브 주소는 이거야: https://github.com/leejunsu0530/studycam

+++ studycam/schedule.py
@@ -55,6 +55,6 @@ class StudyStore:
             cursor -= timedelta(days=1)
         return count
 
-    def save_settings(self, study: int, rest: int, capture: int, speed: int) -> None:
+    def save_settings(self, study: int, rest: int, capture: float, speed: int) -> None:
         self.data["settings"] = {"study_minutes": study, "break_minutes": rest, "capture_seconds": capture, "speed": speed}
         self.save()
